Bump v0.8.2-DIY & rotary/robustness fixes
Update project version to v0.8.2-DIY and add release UF2 artifacts; remove an outdated build archive. Add release notes to README describing rotary encoder behavior changes and fixes. Major fixes and hardening across firmware: enforce NEOPICO_MAX_LEDS and bound LED buffers/numPixels; widen chase index and various frame arrays. Improve rotary encoder API and safety: introduce preprocess/reinit/DrainPendingCounts, track IRQ registrations, clarify QDEC LUT, and add diagnostics state. Strengthen concurrency and event handling: add handler owner tokens to EventManager, GPEncoderEvent magnitude, storagemanager publish/snapshot API and mutex, and mark gp2040aux isReady volatile. Numerous robustness fixes: guard divide-by-zero and invalid inputs (analog deadzone, buzzer/rumble frequency and song lengths, HE trigger mux, PWM frequency), null/length checks and safer memcmp in USB host listeners, rollover-safe polling, display event registration with owner and error-screen handling, input-macro bounds checks, keyboard volume accumulator widened to int32, and assorted defensive comments/clarifications. These changes reduce crashes, prevent UB, and make encoder/LED/host handling more reliable.

+++ headers/drivers/keyboard/KeyboardDriver.h
@@ -35,7 +35,11 @@ private:
     uint8_t last_report[CFG_TUD_ENDPOINT0_SIZE] = { };
     uint16_t last_report_size;
     KeyboardReport keyboardReport;
-    int8_t volumeChange;
+    // Was int8_t. Rotary-encoder volume events arrive with magnitude up to 16 per event,
+    // and the per-tick HID send only decrements this by 1. Fast spinning could drive an
+    // int8_t accumulator past +/-127, which is signed-overflow UB. Use int32_t and saturate
+    // explicitly in handleEncoder() so the driver behaviour stays well-defined.
+    int32_t volumeChange;
 };
 
 #endif // _KEYBOARD_DRIVER_H_
